refactor genesis related codes
* Remove unused MultiChainId and MultiTransactionPayload types
* move StdlibVersion to starcoin-types
* move vm2 ChainId to on_chain_resource
* remove vm2 StdlibVersion/ConsensusStrategy
* move vm2 GenesisConfig to starcoin-config
