+++ executor/tests/module_upgrade_test.rs
@@ -11,12 +11,13 @@ use starcoin_types::account_config::config_change::ConfigChangeEvent;
 use starcoin_types::account_config::TwoPhaseUpgradeV2Resource;
 use starcoin_types::identifier::Identifier;
 use starcoin_types::language_storage::{ModuleId, StructTag, TypeTag};
+use starcoin_types::stdlib::StdlibVersion;
 use starcoin_types::transaction::ScriptFunction;
 use starcoin_vm_types::access_path::AccessPath;
 use starcoin_vm_types::account_config::upgrade::UpgradeEvent;
 use starcoin_vm_types::account_config::{association_address, core_code_address, AccountResource};
 use starcoin_vm_types::account_config::{genesis_address, stc_type_tag};
-use starcoin_vm_types::genesis_config::{ChainId, StdlibVersion};
+use starcoin_vm_types::genesis_config::ChainId;
 use starcoin_vm_types::move_resource::MoveResource;
 use starcoin_vm_types::on_chain_config::{MoveLanguageVersion, TransactionPublishOption, Version};
 use starcoin_vm_types::on_chain_resource::LinearWithdrawCapability;
