+++ vm/types/src/genesis_config.rs
@@ -6,90 +6,9 @@ use anyhow::{format_err, Result};
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 use schemars::{self, JsonSchema};
 use serde::{Deserialize, Serialize};
-use std::cmp::Ordering;
 use std::fmt::{self, Formatter};
 use std::fmt::{Debug, Display};
 use std::str::FromStr;
-#[derive(Clone, Copy, Debug, Default, Deserialize, Eq, Hash, PartialEq, Serialize)]
-pub enum StdlibVersion {
-    #[default]
-    Latest,
-    Version(VersionNumber),
-}
-
-type VersionNumber = u64;
-
-impl StdlibVersion {
-    pub fn new(version: u64) -> Self {
-        if version == 0 {
-            StdlibVersion::Latest
-        } else {
-            StdlibVersion::Version(version)
-        }
-    }
-
-    pub fn as_string(&self) -> String {
-        match self {
-            StdlibVersion::Latest => "latest".to_string(),
-            StdlibVersion::Version(version) => format!("{}", version),
-        }
-    }
-
-    pub fn version(&self) -> u64 {
-        match self {
-            StdlibVersion::Latest => 0,
-            StdlibVersion::Version(version) => *version,
-        }
-    }
-
-    pub fn is_latest(&self) -> bool {
-        matches!(self, StdlibVersion::Latest)
-    }
-
-    pub fn compatible_with_previous(version: &StdlibVersion) -> bool {
-        // currently only 4 is not compatible with previous version
-        // Todo: need a better solution
-        !matches!(version, StdlibVersion::Version(4))
-    }
-}
-
-impl PartialOrd for StdlibVersion {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for StdlibVersion {
-    fn cmp(&self, other: &Self) -> Ordering {
-        match (self, other) {
-            (StdlibVersion::Latest, StdlibVersion::Latest) => Ordering::Equal,
-            (StdlibVersion::Latest, _) => Ordering::Greater,
-            (_, StdlibVersion::Latest) => Ordering::Less,
-            (StdlibVersion::Version(self_v), StdlibVersion::Version(other_v)) => {
-                self_v.cmp(other_v)
-            }
-        }
-    }
-}
-impl FromStr for StdlibVersion {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "latest" => Ok(StdlibVersion::Latest),
-            s => Ok(Self::new(s.parse()?)),
-        }
-    }
-}
-
-impl Display for StdlibVersion {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            StdlibVersion::Latest => f.write_str("latest"),
-            StdlibVersion::Version(version) => f.write_str(version.to_string().as_str()),
-        }
-    }
-}
 
 #[derive(
     Clone,
@@ -123,7 +42,7 @@ impl ConsensusStrategy {
     }
 }
 
-impl fmt::Display for ConsensusStrategy {
+impl Display for ConsensusStrategy {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             ConsensusStrategy::Dummy => write!(f, "dummy"),
