+++ vm2/genesis/Cargo.toml
@@ -11,6 +11,8 @@ version.workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
+starcoin-config = { workspace = true }
+starcoin-storage = { workspace = true }
 starcoin-vm2-vm-types = { workspace = true }
 starcoin-vm2-types = { workspace = true }
 starcoin-vm2-crypto = { workspace = true }
@@ -18,4 +20,3 @@ starcoin-vm2-executor = { workspace = true }
 starcoin-vm2-state-api = { workspace = true }
 starcoin-vm2-transaction-builder = { workspace = true }
 starcoin-vm2-statedb = { workspace = true }
-starcoin-storage = { workspace = true }
