+++ vm2/types/src/account.rs
@@ -23,10 +23,10 @@ use starcoin_vm2_vm_types::{
         self, coin_store::CoinStoreResource, core_code_address, stc_type_tag, AccountResource,
     },
     event::EventKey,
-    genesis_config::ChainId,
     identifier::Identifier,
     language_storage::ModuleId,
     language_storage::StructTag,
+    on_chain_resource::ChainId,
     state_store::state_key::StateKey,
     transaction::authenticator::{AccountPrivateKey, AccountPublicKey},
     transaction::EntryFunction,
