+++ vm2/vm-types/src/block_metadata/mod.rs
@@ -6,7 +6,7 @@
 
 use crate::account_address::AccountAddress;
 use crate::account_config::genesis_address;
-use crate::genesis_config::ChainId;
+use crate::on_chain_resource::ChainId;
 use bcs_ext::Sample;
 use serde::{Deserialize, Deserializer, Serialize};
 use starcoin_crypto::hash::PlainCryptoHash;
