+++ vm2/vm-types/src/on_chain_resource/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod block_metadata;
+mod chain_id;
 pub mod dao;
 mod epoch;
 mod global_time;
@@ -9,6 +10,7 @@ pub mod nft;
 mod treasury;
 
 pub use block_metadata::BlockMetadata;
+pub use chain_id::ChainId;
 pub use epoch::{Epoch, EpochData, EpochInfo};
 pub use global_time::GlobalTimeOnChain;
 pub use treasury::{LinearWithdrawCapability, Treasury};
